+++ client/image_list.go
@@ -41,6 +41,12 @@ func (cli *Client) ImageList(ctx context.Context, options ImageListOptions) (Ima
 			query.Set("manifests", "1")
 		}
 	}
+	if options.Identity {
+		if err := cli.requiresVersion(ctx, "1.53", "identity"); err != nil {
+			return ImageListResult{}, err
+		}
+		query.Set("identity", "1")
+	}
 
 	resp, err := cli.get(ctx, "/images/json", query, nil)
 	defer ensureReaderClosed(resp)